+++ public/main.js
@@ -1025,7 +1025,7 @@ function load(saveString, autoLoad, fromPf) {
 				else {
 					map.loot++;
 				}
-				
+
 			}
 		}
 		if (game.global.highestRadonLevelCleared >= 200){
@@ -1540,7 +1540,7 @@ function toggleHideStat(stat, elem){
 	var stat = game.stats[stat];
 	stat.hidden = !stat.hidden;
 	elem.className = "btn btn-" + ((stat.hidden) ? 'danger' : 'success') + " btn-lg";
-	elem.innerHTML = (stat.hidden) ? "Unhide" : "Hide";  
+	elem.innerHTML = (stat.hidden) ? "Unhide" : "Hide";
 }
 
 function countChallengeSquaredReward(numberOnly, mesmerPreview, getUniverseArray){
@@ -1953,14 +1953,14 @@ function abandonChallenge(restart){
 			fadeIn("helium", 10);
 			game.global.runningChallengeSquared = false;
 			if (game.global.universe == 2 && (game.global.world > 30 || (game.global.world == 30 && game.global.lastClearedCell >= 29))) unlockJob("Meteorologist");
-		}	
+		}
 	}
 	else if (challenge.fireAbandon && typeof challenge.abandon !== 'undefined') {
 		game.global.challengeActive = "";
 		challenge.abandon();
 	}
 	game.global.challengeActive = "";
-	
+
 	cancelPortal();
 	if (challengeName == "Scientist"){
 		document.getElementById("scienceCollectBtn").style.display = "block";
@@ -2366,7 +2366,7 @@ function manageEqualityStacks(){
 	var enemyMult = game.portal.Equality.getMult(false);
 	if (game.heirlooms.Shield.inequality.currentBonus > 0){
 		var trimpMult = game.portal.Equality.getMult(true);
-		
+
 		text += " by " + prettifyTiny(trimpMult) + " and Enemies by " + prettifyTiny(enemyMult);
 	}
 	else {
@@ -2398,7 +2398,7 @@ function scaleEqualityScale(slider, whatDo){
 		if (!(val >= 0) || !(val <= 10)) val = 5;
 		game.portal.Equality.scalingSetting = val;
 		textElem = document.getElementById('equalityCurrentScale');
-		
+
 	}
 	else if (slider.id == "scaleEqualitySliderReverse"){
 		if (!(val >= 0) || !(val <= 10)) val = 5;
@@ -2457,7 +2457,7 @@ function updatePerkColor(what){
 		perkClass = ((canSpend >= price)) ? "perkColorOn" : "perkColorOff";
 		if (usingScreenReader){
 			var affordElem = document.getElementById('screenReaderPerkAfford' + what);
-			
+
 			if (affordElem){
 				if (perkClass == "perkColorMaxed") affordElem.innerHTML = ", Max";
 				else affordElem.innerHTML = (removableLevel > 0) ? ", Can Buy" : ", Not Affordable";
@@ -2611,7 +2611,7 @@ var offlineProgress = {
 				this.mapTextElem.innerHTML = "Oof, you don't have enough fragments to run a map."
 			}
 			else this.mapTextElem.innerHTML = "You can run <b>" + this.mapsAllowed + " map" + needAnS(this.mapsAllowed) + "</b> while you wait!<br/>Use " + ((this.mapsAllowed == 1) ? "it" : "them") + " wisely...<br/>You have " + prettify(frags) + " Fragments.";
-			this.mapBtns[x].style.display = (price > frags) ? 'none' : 'inline-block';		
+			this.mapBtns[x].style.display = (price > frags) ? 'none' : 'inline-block';
 			this.mapBtns[x].innerHTML = "Z " + useWorld + " map<br/>" + prettify(price) + " Frags<br/>" + this.countMapItems(useWorld) + " items";
 		}
 	},
@@ -3124,7 +3124,7 @@ function countHeliumSpent(checkTemp, usePortalUniverse){
 			if (typeof item.level === 'undefined') continue;
 			count += item.heliumSpent;
 		}
-		
+
 		if (checkTemp && item.heliumSpentTemp) count += item.heliumSpentTemp;
 	}
 	return count;
@@ -3215,7 +3215,7 @@ function getPerkBuyCount(perkName){
 		toBuy--;
 	}
 	else {
-		//Let it be known that buy max on additive perks only exists because The Amazing GhostFrog is a math wizard. 
+		//Let it be known that buy max on additive perks only exists because The Amazing GhostFrog is a math wizard.
 		//I spent hours bashing my head against a wall trying to figure it out, and she came up with this formula in a few minutes.
 		var levels = perkLevels + perk.levelTemp;
 		var increase = perk.additiveInc
@@ -3939,7 +3939,7 @@ function rewardResource(what, baseAmt, level, checkMapLootScale, givePercentage)
 
 	}
 	//Add Looting
-	
+
 	if (getPerkLevel("Looting")) amt += (amt * getPerkLevel("Looting") * game.portal.Looting.modifier);
 	if (getPerkLevel("Looting_II")) amt *= (1 + (getPerkLevel("Looting_II") * game.portal.Looting_II.modifier));
 	if (what == "helium") amt *= alchObj.getRadonMult();
@@ -4725,7 +4725,7 @@ function trapThings() {
 		var trapValue = 1;
 		if (getPerkLevel("Bait") > 0) trapValue += (getPerkLevel("Bait") * game.portal.Bait.modifier);
 		if (Fluffy.isRewardActive("trapper")) trapValue *= 10;
-        trimps.owned += trapValue;		
+        trimps.owned += trapValue;
 		if (trimps.owned > trimpsMax) trimps.owned = trimpsMax;
         game.global.timeLeftOnTrap = -1;
 		if (TrapOwned) TrapOwned.innerHTML = trap.owned;
@@ -4801,7 +4801,7 @@ function buyAutoJobs(allowRatios){
 			console.log('wanted to buy ' + toBuy + ' ' + ratios[x] + 's.');
 			continue;
 		}
-		var useMax = (ratios[x] == "Scientist" && thisSetting.buyMax) ? thisSetting.buyMax : false; 
+		var useMax = (ratios[x] == "Scientist" && thisSetting.buyMax) ? thisSetting.buyMax : false;
 		autoBuyJob(item, true, Math.floor(toBuy), useMax);
 	}
 	calculateParityBonus();
@@ -4846,7 +4846,7 @@ function autoBuyJob(what, isRatio, purchaseAmt, max){
 			if (workspaces < buyAmt && freed){
 				workspaces = buyAmt;
 				checkAndFix = true;
-			}	
+			}
 		}
 	}
 	for (var costItem in job.cost) {
@@ -5217,7 +5217,7 @@ var srLastBreedTime = "";
 function breed() {
     var trimps = game.resources.trimps;
 	checkAchieve("trimps", trimps.owned);
-    
+
 	var trimpsMax = trimps.realMax();
 	var employedTrimps = trimps.employed;
 	if (game.permaBoneBonuses.multitasking.owned) employedTrimps *= (1 - game.permaBoneBonuses.multitasking.mult());
@@ -5345,10 +5345,10 @@ function breed() {
 				swapClass("state", "stateHappy", GAElem);
 				GAIndicator.innerHTML = "";
 			}
-			
+
 		}
 	}
-	
+
 	timeRemaining = timeRemaining.toNumber();
 	totalTime = totalTime.toNumber();
 	decimalOwned = decimalOwned.add(breeding.div(10));
@@ -5701,14 +5701,14 @@ function resetAdvMaps(fromClick) {
 	}
 	//biome
 	var biomeElem = document.getElementById("biomeAdvMapsSelect");
-	if (game.global.decayDone && document.getElementById('gardenOption') === null) 
+	if (game.global.decayDone && document.getElementById('gardenOption') === null)
 		biomeElem.innerHTML += "<option id='gardenOption' value='Plentiful'>Gardens</option>";
-	if (game.global.farmlandsUnlocked && document.getElementById('farmlandsOption') === null) 
+	if (game.global.farmlandsUnlocked && document.getElementById('farmlandsOption') === null)
 		biomeElem.innerHTML += "<option id='farmlandsOption' value='Farmlands'>Farmlands</option>";
 	biomeElem.value = (preset.biome && !fromClick) ? preset.biome : "Random";
 	//bottom row
 	hideAdvMaps(true);
-	
+
 	document.getElementById('advSpecialSelect').value = (!fromClick && preset.specMod) ? preset.specMod : "0";
 	swapNiceCheckbox(document.getElementById('advPerfectCheckbox'), (!fromClick && preset.perf));
 	document.getElementById('advExtraLevelSelect').value = (!fromClick && preset.extra > 0) ? preset.extra.toString() : "0";
@@ -5909,7 +5909,7 @@ function adjustMap(what, value) {
 	var text = "";
 	if (checkPerfectChecked())
 		text = (what == "loot") ? minMax[1] : minMax[0];
-	else 
+	else
 		text = "Min " + minMax[0] + ", Max " + minMax[1];
 	document.getElementById(what + "AdvMapsText").innerHTML = text;
 	updateMapCost();
@@ -6293,7 +6293,7 @@ function createVoidMap(forcePrefix, forceSuffix, skipMessage, noDupe) {
 			stacked: 0
 		});
 		if (game.talents.voidPower3.purchased)
-			map.bonus = 'fa';	
+			map.bonus = 'fa';
 		game.global.mapsOwnedArray.push(map);
 	}
 	game.global.totalVoidMaps++;
@@ -6603,7 +6603,7 @@ function displayAddCarriedBtn(){
 }
 
 //RIP buying carried slots
-// function addCarried(confirmed){ 
+// function addCarried(confirmed){
 // 	if (game.global.maxCarriedHeirlooms > game.heirlooms.values.length){
 // 		elem.style.display = "none";
 // 		return;
@@ -6831,7 +6831,7 @@ function toggleHeirloomOnPortal(){
 	var equipped = game.global[type + "Equipped"];
 	if (on && id == equipped.id) equipped.onPortal = true;
 	else if (equipped) equipped.onPortal = false;
-	
+
 	for (var x = 0; x < game.global.heirloomsCarried.length; x++){
 		var heirloom = game.global.heirloomsCarried[x];
 		if (on && heirloom.id == id) heirloom.onPortal = true;
@@ -7615,7 +7615,7 @@ function getHeirloomRarityRanges(zone, forBones, forceIndex){
 		if (game.global.totalRadPortals > 0) zone = game.global.highestRadonLevelCleared + 1;
 		else zone = game.global.highestLevelCleared + 1;
 	}
-	var useBreakpoint = (typeof forceIndex !== 'undefined') ? forceIndex : getHeirloomZoneBreakpoint(zone, forBones); 
+	var useBreakpoint = (typeof forceIndex !== 'undefined') ? forceIndex : getHeirloomZoneBreakpoint(zone, forBones);
 	var rarities = game.heirlooms.rarities[useBreakpoint];
 	var canLower = 0;
 	var addBonus = false;
@@ -8130,7 +8130,7 @@ var mutations = {
 				currentArray[last] = "Living";
 			}
 			if (!fromPattern) this.updateGrid(currentArray);
-			this.savePattern(currentArray, true);	
+			this.savePattern(currentArray, true);
 			return currentArray;
 		},
 		checkDirection: function (amt, x){
@@ -8191,7 +8191,7 @@ var mutations = {
 			var arrayHolder = game.challenges.Life.arrayHolder;
 			for (var x = 0; x < arrayHolder.length - 1; x++){
 				var comparing = arrayHolder[x];
-				mid: 
+				mid:
 				for (var y = x + 1; y < arrayHolder.length; y++) {
 					var compareTo = arrayHolder[y];
 					if (comparing.length != compareTo.length) continue;
@@ -8234,7 +8234,7 @@ var mutations = {
 							document.getElementById('livingMutationName').innerHTML = "";
 						}
 					}
-				}			
+				}
 			}
 		},
 		nextMove: function (currentArray){
@@ -8733,10 +8733,10 @@ var mutations = {
 		cellCount: function(){
 			var lastSpire = game.global.lastSpireCleared;
 			var world = game.global.world;
-			if (lastSpire < 2) 
+			if (lastSpire < 2)
 				return 0;
 			lastSpire *= 100;
-			if (world > lastSpire + 199) 
+			if (world > lastSpire + 199)
 				world = lastSpire + 199;
 			var possible = Math.floor((world - 300) / 15) + 2;
 			if (game.talents.healthStrength2.purchased) possible += game.global.lastSpireCleared;
@@ -8939,7 +8939,7 @@ var visualMutations = {
 		},
 		namePrefix: "Snowy"
 	},
-	
+
 }
 
 function startTheMagma(){
@@ -9128,7 +9128,7 @@ function checkGenStateSwitch(){
 			return;
 		}
 	}
-	
+
 }
 
 function getGeneratorHtml(getContainer){
@@ -9163,7 +9163,7 @@ function updatePauseBtn(show){
 	else {
 		updateNextGeneratorTickTime();
 	}
-	
+
 }
 
 function pauseGenerator(updateOnly){
@@ -9809,7 +9809,7 @@ function addSpecials(maps, countOnly, map, getPrestiges) { //countOnly must incl
 		if ((special.world == -10) && ((world % 10) !== 0)) continue;
 		if ((special.world == -20) && ((world % 20) !== 0)) continue;
 		if ((special.world == -25) && ((world % 25) !== 0)) continue;
-		
+
 		if ((maps) && (special.filter)){
 			var mapResType = game.mapConfig.locations[map.location].resourceType;
 			if (mapResType == "Scaling") mapResType = getFarmlandsResType(map);
@@ -9848,7 +9848,7 @@ function addSpecials(maps, countOnly, map, getPrestiges) { //countOnly must incl
 					addSpecialToNthLast(special, array, item, 3);
 					canLast = 2;
 				}
-				continue;	
+				continue;
 			}
 			array = addSpecialToLast(special, array, item);
             canLast = 0;
@@ -9986,7 +9986,7 @@ function drawGrid(maps) { //maps t or f. This function overwrites the current gr
 	}
 	var className = "";
 	if (game.global.universe == 1 && !maps && game.global.world >= 60 && game.global.world <= 80){
-		
+
 		if (game.global.world == 60) className = "gridOverlayGreenGradient1"
 		else if (game.global.world <= 65) className = "gridOverlayGreenGradient2";
 		else if (game.global.world <= 70) className = "gridOverlayGreenGradient3";
@@ -10023,7 +10023,7 @@ function drawGrid(maps) { //maps t or f. This function overwrites the current gr
 				cell.style.background = "initial";
 				cell.style.backgroundColor = u2Mutations.getColor(game.global.gridArray[counter].u2Mutation);
 				className += " mutatedCell";
-				
+
 			}
 			else if (!maps && game.global.gridArray[counter].mutation) className += " " + game.global.gridArray[counter].mutation;
 			if (!maps && game.global.gridArray[counter].vm){
@@ -10286,7 +10286,7 @@ function mapsSwitch(updateOnly, fromRecycle) {
         if (game.global.preMapsActive) {
             game.global.mapsActive = false;
             game.global.preMapsActive = false;
-		} 
+		}
 		else game.global.preMapsActive = true;
 	}
 	if (!updateOnly)
@@ -10322,7 +10322,7 @@ function mapsSwitch(updateOnly, fromRecycle) {
         if (game.global.lookingAtMap && !game.global.currentMapId) selectMap(game.global.lookingAtMap, true);
 		else if (game.global.currentMapId === "") {
 			clearMapDescription();
-		} 
+		}
 		else {
             selectMap(game.global.currentMapId, true);
             document.getElementById("selectMapBtn").innerHTML = "Continue";
@@ -10358,7 +10358,7 @@ function mapsSwitch(updateOnly, fromRecycle) {
 		u2Mutations.types.Nova.clearStacks();
 		u2Mutations.types.Rage.clearStacks();
         toggleMapGridHtml(true, currentMapObj);
-	} 
+	}
 	else {
 		//Switching to world
 		if (!updateOnly) resetEmpowerStacks();
@@ -10423,7 +10423,7 @@ function setNonMapBox(){
 	if (game.talents.mapBattery.purchased && bonus == 10) bonus *= 2;
 	if (bonus > 0) mapBonus.innerHTML = prettify(bonus * 20) + "% Map Bonus";
 	else mapBonus.innerHTML = "";
-	document.getElementById("worldName").innerHTML = (game.global.spireActive) ? ((checkIfSpireWorld(true) == 1) ? "Spire" : "Spire " + romanNumeral(checkIfSpireWorld(true))) : "Zone";	
+	document.getElementById("worldName").innerHTML = (game.global.spireActive) ? ((checkIfSpireWorld(true) == 1) ? "Spire" : "Spire " + romanNumeral(checkIfSpireWorld(true))) : "Zone";
 }
 
 function repeatClicked(updateOnly){
@@ -10435,7 +10435,7 @@ function repeatClicked(updateOnly){
 	elem.className = "btn fightBtn " + color;
 	elem.innerHTML = (game.global.repeatMap) ? "Repeat On" : "Repeat Off";
 	if (elem2 !== null){
-		color = (game.global.repeatMap) ? "settingBtn1" : "settingBtn0"; 
+		color = (game.global.repeatMap) ? "settingBtn1" : "settingBtn0";
 		swapClass("settingBtn", color, elem2);
 		elem2.innerHTML = (game.global.repeatMap) ? "Repeat On" : "Repeat Off";
 	}
@@ -10505,7 +10505,7 @@ function runMap() {
     if (game.global.lastClearedMapCell == -1) {
         buildMapGrid(mapId);
         drawGrid(true);
-		
+
 		if (mapObj.location == "Void"){
 			game.global.voidDeaths = 0;
 			game.global.voidBuff = mapObj.voidBuff;
@@ -10603,14 +10603,14 @@ function battle(force) {
 function checkAmalgamate(){
 	if (game.global.universe == 2) return;
 	var amalgJoinTexts = [
-		"A small black hole opens up in the sky and a shrill noise echoes across the town. A few moments later, an Amalgamator is standing in front of you. It looks around briefly, grabs a few Trimps, and runs off towards your soldiers.", 
+		"A small black hole opens up in the sky and a shrill noise echoes across the town. A few moments later, an Amalgamator is standing in front of you. It looks around briefly, grabs a few Trimps, and runs off towards your soldiers.",
 		"While out walking a Trimp, you suddenly find yourself teleported back to your ship, standing face to face with an Amalgamator. You introduce yourself but it doesn't seem up for conversation.",
 		"Suddenly, the largest rain drops you've ever seen start falling from the sky - each drop is at least 1000 times larger than normal. One particularly large drop hits the ground and an Amalgamator pops out! It sends out a quick telepathic greeting, then goes off to find your Trimps.",
 		"You're sitting down about to enjoy a rare dinner break, when an Amalgamator gets interested in your dimension and replaces the spacetime of your meal with itself. You really hope they don't do that again.",
 		"As you're helping your Trimps cross a deeper-than-average stream, you notice a column of bubbles coming up near your Trimps. A gurgling sound begins to grow from the source of the bubbles, and your Trimps start to get a little freaked out. Suddenly an Amalgamator bursts to the surface, spits some water at a Trimp, then teleports to your town."];
 	var amalgLeaveTexts = [
-		"You hear the second or third most chilling sound you've ever heard, and your Amalgamator pops up in front of you. It smacks you with a small stick to show dissatisfaction, then disappears in a puff of smoke.", 
-		"While walking through your town, you notice your Amalgamator throwing a fit, kicking over food carts and anything else not tied down. You approach it but before you can ask what's wrong, it smacks you with a small stick to show dissatisfaction, then it scurries away.", 
+		"You hear the second or third most chilling sound you've ever heard, and your Amalgamator pops up in front of you. It smacks you with a small stick to show dissatisfaction, then disappears in a puff of smoke.",
+		"While walking through your town, you notice your Amalgamator throwing a fit, kicking over food carts and anything else not tied down. You approach it but before you can ask what's wrong, it smacks you with a small stick to show dissatisfaction, then it scurries away.",
 		"You watch as your Amalgamator struggles to find enough free Trimps, panic searching in places such as under rocks or between the leaves of trees. It suddenly seems to remember that it doesn't have to be there, smacks you with a small stick to show dissatisfaction, and turns into nothing.",
 		"While in town, a Scientist approaches you to let you know that your Amalgamator is getting upset and to keep an eye out for him. Just as you're finishing the conversation, the Amalgamator appears in front of you. It smacks you both with a small stick to show dissatisfaction, then turns into a small puddle of water - which you ask the Scientist to clean up."];
 	if (game.global.challengeActive == "Trapper" || game.global.challengeActive == "Trappapalooza") return false;
@@ -11041,7 +11041,7 @@ function startFight() {
 			instaFight = true;
 			if (!game.global.mapsActive) game.stats.cellsOverkilled.value++;
 		}
-		else 
+		else
 		{
 			if (cell.plaguebringer){
 				if (cell.health > cell.maxHealth * 0.05){
@@ -11146,7 +11146,7 @@ function startFight() {
 		}
 		//Soldiers
 		game.global.soldierHealthMax *= trimpsFighting;
-		game.global.soldierCurrentAttack *= trimpsFighting;	
+		game.global.soldierCurrentAttack *= trimpsFighting;
 		//Toughness
 		if (getPerkLevel("Toughness") > 0) game.global.soldierHealthMax += (game.global.soldierHealthMax * getPerkLevel("Toughness") * game.portal.Toughness.modifier);
 		if (getPerkLevel("Toughness_II") > 0) game.global.soldierHealthMax *= (1 + (game.portal.Toughness_II.modifier * getPerkLevel("Toughness_II")));
@@ -11219,7 +11219,7 @@ function startFight() {
 			game.global.soldierHealthMax *= game.challenges.Life.getHealthMult();
 		}
 		if (game.global.challengeActive == "Revenge") game.global.soldierHealthMax *= game.challenges.Revenge.getMult();
-		if (game.global.challengeActive == "Duel" && game.challenges.Duel.trimpStacks < 20) game.global.soldierHealthMax *= game.challenges.Duel.healthMult;	
+		if (game.global.challengeActive == "Duel" && game.challenges.Duel.trimpStacks < 20) game.global.soldierHealthMax *= game.challenges.Duel.healthMult;
 		if (game.talents.voidPower.purchased && game.global.voidBuff){
 			game.global.soldierHealthMax *= ((game.talents.voidPower.getTotalVP() / 100) + 1);
 			game.global.voidPowerActive = true;
@@ -11278,7 +11278,7 @@ function startFight() {
 				game.global.soldierHealthMax /= 2;
 				if (game.global.soldierHealth > game.global.soldierHealthmax) game.global.soldierHealth = game.global.soldierHealthMax;
 				game.global.mapHealthActive = false;
-				if (game.global.universe == 2){ 
+				if (game.global.universe == 2){
 					game.global.soldierEnergyShieldMax /= 2;
 					if (game.global.soldierEnergyShield > game.global.soldierEnergyShieldMax) game.global.soldierEnergyShield = game.global.soldierEnergyShieldMax;
 				}
@@ -11518,7 +11518,7 @@ function getEnergyShieldMult(){
 	if (game.global.challengeActive == "Bublé") total += 2.5; //Bublé challenge - 100%
 	if (autoBattle.oneTimers.Suprism.owned) total += autoBattle.oneTimers.Suprism.getMult();
 	if (getHeirloomBonus("Shield", "prismatic") > 0) total += (getHeirloomBonus("Shield", "prismatic") / 100);
-	
+
 	return total;
 }
 
@@ -11671,7 +11671,7 @@ function calculateDamage(number, buildString, isTrimp, noCheckAchieve, cell, noF
 		}
 		if (autoBattle.bonuses.Stats.level > 0 && game.global.universe == 2) number *= autoBattle.bonuses.Stats.getMult();
 		if (game.global.challengeActive == "Alchemy") number *= alchObj.getPotionEffect("Potion of Strength");
-		
+
 		if (game.global.pandCompletions){
 			number *= game.challenges.Pandemonium.getTrimpMult();
 		}
@@ -11971,7 +11971,7 @@ function displayTalents(){
 			if (Array.isArray(talent.requires)) requires = talent.requires;
 			else requires = [talent.requires];
 			for (var x = 0; x < requires.length; x++){
-				if (!game.talents[requires[x]].purchased){ 
+				if (!game.talents[requires[x]].purchased){
 					talentClass += " talentReqNeeded";
 					break;
 				}
@@ -11990,7 +11990,7 @@ function displayTalents(){
 			else if (talentClass.search('NotPurchased') >= 0) statusText = "Not Purchased";
 			else statusText = "Purchased"
 			html += "<div role='button' class='" + talentClass + "' onclick='tooltip(\"" + item + "\", \"talents\", \"screenRead\")'>" + talent.name + " Info</div>";
-			html += "<div role='button' class='" + talentClass + "' onclick='purchaseTalent(\"" + item + "\")'><span class='talentIcon'><span class='" + icon + "'></span></span><br/><div class='talentName'>Tier " + talent.tier + " " + statusText + " " + talent.name + "</div></div>";			
+			html += "<div role='button' class='" + talentClass + "' onclick='purchaseTalent(\"" + item + "\")'><span class='talentIcon'><span class='" + icon + "'></span></span><br/><div class='talentName'>Tier " + talent.tier + " " + statusText + " " + talent.name + "</div></div>";
 		}
 		else {
 			html += "<div class='" + talentClass + "' onmouseover='tooltip(\"" + item + "\", \"talents\", event)' onmouseout='tooltip(\"hide\")' onclick='purchaseTalent(\"" + item + "\")'><span class='talentIcon'><span class='" + icon + "'></span></span><br/><div class='talentName'>" + talent.name + "</div></div>";
@@ -12091,7 +12091,7 @@ function completeTalentPurchase(talent){
 		if (Array.isArray(talent.requires)) requires = talent.requires;
 		else requires = [talent.requires];
 		for (var x = 0; x < requires.length; x++){
-			if (!game.talents[requires[x]].purchased){ 
+			if (!game.talents[requires[x]].purchased){
 				return;
 			}
 		}
@@ -12121,7 +12121,7 @@ function canPurchaseRow(tierNumber){
 		var purchasedThisTier = countPurchasedTalents(x);
 		totalRequiredPurchases += (6 - purchasedThisTier);
 	}
-	if (checkAffordableTalents() - countPurchasedTalents() >= totalRequiredPurchases) 
+	if (checkAffordableTalents() - countPurchasedTalents() >= totalRequiredPurchases)
 		return true;
 	return false;
 }
@@ -12253,7 +12253,7 @@ function rewardLiquidZone(){
 		if (checkVoidMap() == 1) voidMaps++;
 		var cell = game.global.gridArray[x];
 		if (cell.special !== ""){
-			var unlock = game.worldUnlocks[cell.special];        
+			var unlock = game.worldUnlocks[cell.special];
 			if (typeof unlock !== 'undefined' && typeof unlock.fire !== 'undefined') {
 				unlock.fire(x);
 				if (hiddenUpgrades.indexOf(cell.special) == -1){
@@ -12262,7 +12262,7 @@ function rewardLiquidZone(){
 					if (typeof unlock.displayAs !== 'undefined')
 						unlocks[index] += unlock.displayAs;
 					else
-						unlocks[index] += cell.special;	
+						unlocks[index] += cell.special;
 				}
 			} else {
 				unlockEquipment(cell.special);
@@ -12309,7 +12309,7 @@ function rewardLiquidZone(){
 			text += " Wood - " + prettify(game.resources.wood.owned - wood) + ",";
 			text += " Metal - " + prettify(game.resources.metal.owned - metal) + ",";
 		}
-		
+
 		text = text.slice(0, -1);
 		text += "<br/>";
 	}
@@ -12467,7 +12467,7 @@ function nextWorld() {
 	if (getPerkLevel("Tenacity")){
 		if (game.portal.Tenacity.timeLastZone != -1) game.portal.Tenacity.timeLastZone *= game.portal.Tenacity.getCarryoverMult();
 		game.portal.Tenacity.timeLastZone += getZoneMinutes();
-		
+
 	}
 	game.global.zoneStarted = getGameTime();
 	if (challengeActive("Mapology")) {
@@ -12738,12 +12738,12 @@ function runMapAtZone(index){
 			if (setting.until == 5){ //climbing
 				game.global.mazBw = setting.bwWorld;
 				//if repeating on zones
-				if ((setting.times > 0 || setting.times == -2) && game.global.world > setting.world){			
+				if ((setting.times > 0 || setting.times == -2) && game.global.world > setting.world){
 					//see how many times this has repeated by zone, increase target climb level by appropriate amount for zones skipped
 					var times = (setting.times == -2) ? setting.tx : setting.times;
 					var repeats = Math.round((game.global.world - setting.world) / times);
 					if (repeats > 0) game.global.mazBw += (times * repeats);
-				}					
+				}
 				game.options.menu.repeatUntil.enabled = 2;
 				if (usingRealTimeOffline) offlineProgress.repeatUntil = game.options.menu.repeatUntil.enabled;
 				toggleSetting('repeatUntil', null, false, true);
@@ -13049,8 +13049,8 @@ function getSpireStory(spireNum, row){
 		},
 		spire5: {
 			r2: "While you walk through the Spire, you whistle a tune. You haven't yet tired, your motivations zoom. As Fluffy checks for loot in this dark, dusty room, you suddenly see fire and hear a huge boom. ",
-			r6: "Against your better judgment, you continue to climb. Heat fills this dungeon, and doubt fills your mind. You feel a great evil that's long been confined, but with reluctance you move on for the good of Trimpkind. ",			
-			r4: "The fires still grow, but they leave a clear path. Just one direction shown, the thought makes you laugh. Unwisely Druopitee tried to rig this booby trap, but the fire shows the way to go so you thank him for the map. ",			
+			r6: "Against your better judgment, you continue to climb. Heat fills this dungeon, and doubt fills your mind. You feel a great evil that's long been confined, but with reluctance you move on for the good of Trimpkind. ",
+			r4: "The fires still grow, but they leave a clear path. Just one direction shown, the thought makes you laugh. Unwisely Druopitee tried to rig this booby trap, but the fire shows the way to go so you thank him for the map. ",
 			r7: "You're close to the top of this large, awful tower. Your fearful thoughts drop, your will is empowered. Luckily you have Fluffy to support you in this hour, you two will never stop until these echoes are devoured.",
 			r8: "The presence here is unique from what you felt before. You know it can't speak but it's not done for. An intrusive thought knocks on your mind's back door:<br/><span class='spirePoem'>You are weak. I'll be on the top floor.</span> ",
 			r9: "Another thought enters your mind, this time louder than before:<br/><span class='spirePoem'>Your death warrant's signed, soon you'll be no more.<br/>I gave you everything you have, it was no easy chore.<br/>Your Trimps are MINE, prepare for war!</span><br/>Even in death this guy's a dick. ",
@@ -13115,7 +13115,7 @@ function giveSpireReward(level){
 		case(50):
 			if (spireWorld == 2){
 				if (Fluffy.getCapableLevel() > 0){
-					message("Out of the corner of your eye, you notice a creature who doesn't seem like he belongs. You walk up closer and notice that it's Fluffy! You turn around expecting to see your other Fluffy, but there is only one. You turn back as he waves at you in recognition, you feel a bit lightheaded, but you've had weirder things happen on this planet. You decide to keep moving up the Spire with your new old friend.", "Story");					
+					message("Out of the corner of your eye, you notice a creature who doesn't seem like he belongs. You walk up closer and notice that it's Fluffy! You turn around expecting to see your other Fluffy, but there is only one. You turn back as he waves at you in recognition, you feel a bit lightheaded, but you've had weirder things happen on this planet. You decide to keep moving up the Spire with your new old friend.", "Story");
 				}
 				else {
 					message("Out of the corner of your eye, you notice a creature who doesn't seem like he belongs. You walk up closer and notice that it's a Trimp! Only this Trimp is purple instead of blue, obviously a result of some experiments by Druopitee. You let him know not to worry, and that you'll keep him safe. You name him Fluffy, and vow to never let any harm come to him. <b>Gained 1 Trimp!</b>", "Story");
@@ -13136,7 +13136,7 @@ function giveSpireReward(level){
 			game.global.nullifium += nuReward;
 			message(getSpireStory(spireWorld, 6) + "You found " + nuReward + " Nullifium!</span>", "Story");
 			break;
-		case(70):			
+		case(70):
 			var heirloomLevel = 400;
 			if (spireWorld > 3) heirloomLevel = ((spireWorld + 1) * 100);
 			message(getSpireStory(spireWorld, 7) + "You found a Z" + heirloomLevel + " Heirloom!", "Story");
@@ -13189,7 +13189,7 @@ function giveSpireReward(level){
 				updateBones();
 				text += " For your first time killing this Echo in any timeline, you have permanently added another 5% to your Liquification bonus, increased your Dark Essence gains by 4x, and earned <b>20 bones!</b>"
 				switch (spireWorld){
-					case 2: 
+					case 2:
 						text += " You have also unlocked the <b>Capable</b> perk and can carry an additional Heirloom!";
 						game.portal.Capable.locked = false;
 						break;
@@ -13224,7 +13224,7 @@ function giveSpireReward(level){
 			amt *= Math.pow(1.01, level);
 			amt = giveHeliumReward(amt);
 			message("As you're leaving, you find " + prettify(amt) + " helium!", "Loot", heliumIcon(true), "helium", "helium");
-			return;				
+			return;
 	}
 }
 
@@ -14461,7 +14461,7 @@ function getDailyRewardText(portalUni){
 		if (game.global.highestLevelCleared >= 179) returnText += ((game.global.highestLevelCleared >= 235) ? ", " : ", and ") + "Dark Essence";
 		if (game.global.highestLevelCleared >= 235) returnText += ((game.portal.Capable.locked == false) ? ", " : ", and ") + "Nature Tokens";
 		if (game.portal.Capable.locked == false) returnText += ", and Fluffy Exp";
-	} 
+	}
 	return returnText;
 }
 
@@ -14561,7 +14561,7 @@ function fight(makeUp) {
 		if (game.jobs.Worshipper.owned > 0 && !game.global.mapsActive) tryWorship();
 		if (challengeActive("Nom") && cell.nomStacks == 100) giveSingleAchieve("Great Host");
 		if (game.global.challengeActive == "Obliterated") giveSingleAchieve("Obliterate");
-		if (game.global.challengeActive == "Eradicated") giveSingleAchieve("Eradicate");	
+		if (game.global.challengeActive == "Eradicated") giveSingleAchieve("Eradicate");
 		if (game.global.usingShriek) disableShriek();
 		if (game.global.universe == 2) u2Mutations.types.Rage.clearStacks();
 		//Death message
@@ -15269,7 +15269,7 @@ function fight(makeUp) {
 			else nextCell.plaguebringer += plaguebringer;
 			if (!nextCell.plagueHits) nextCell.plagueHits = getPlaguebringerModifier();
 			else nextCell.plagueHits += getPlaguebringerModifier();
-		}		
+		}
 	}
 	if (game.global.challengeActive == "Devastation" && impOverkill){
 		game.challenges.Devastation.lastOverkill = impOverkill;
@@ -15412,7 +15412,7 @@ function getCritText(critTier){
 	switch(critTier){
 		case -1:
 			return "<span style='color: cyan'>Weak!</span>";
-		case 1: 
+		case 1:
 			return "Crit!";
 		case 2:
 			return "<span style='color: orange'>CRIT!</span>";
@@ -15513,7 +15513,7 @@ function updateLivingStacks(){
 	var stackCount = game.challenges.Life.stacks;
 	if (stackCount < game.challenges.Life.lowestStacks) game.challenges.Life.lowestStacks = stackCount;
 	if (elem === null) {
-		document.getElementById("goodGuyName").innerHTML += '&nbsp<span class="badge antiBadge" id="livingBuff" onmouseover="tooltip(\'Unliving\', null, event)" onmouseout="tooltip(\'hide\')"><span id="livingStacks">' + stackCount + '</span>&nbsp;<span style="margin-top: 2%" class="icomoon icon-shareable"></span></span>';		
+		document.getElementById("goodGuyName").innerHTML += '&nbsp<span class="badge antiBadge" id="livingBuff" onmouseover="tooltip(\'Unliving\', null, event)" onmouseout="tooltip(\'hide\')"><span id="livingStacks">' + stackCount + '</span>&nbsp;<span style="margin-top: 2%" class="icomoon icon-shareable"></span></span>';
 		return;
 	}
 	document.getElementById("livingStacks").innerHTML = stackCount;
@@ -16341,7 +16341,7 @@ function displaySingleRunBonuses(){
 			else{
 				if (game.global.b < bonus.cost)
 					btnClass = 'boneBtnStateOff'
-				else 
+				else
 					btnClass = 'boneBtnStateOn';
 				btnText = bonus.name + " (" + bonus.cost + " bones)";
 			}
@@ -16382,7 +16382,7 @@ function displayPermaBoneBonuses(){
 			desc = bonus.text;
 			if (game.global.b < cost)
 				btnClass = 'boneBtnStateOff'
-			else 
+			else
 				btnClass = 'boneBtnStateOn';
 				btnText = bonus.name + " (" + bonus.owned +  "/10, " + cost + " Bones)";
 		}
@@ -16590,7 +16590,7 @@ var sugarRush = {
 	enableIcon: function () {
 		var elem = this.getIconElement();
 		if (!elem){
-			document.getElementById('goodGuyName').innerHTML += ' <span class="badge antiBadge sugarRushBadge" id="sugarRushBuff" onmouseover="tooltip(\'Sugar Rush\', \'customText\', event, sugarRush.tooltipText())" onmouseout="tooltip(\'hide\')"><span class="' + this.icon + '"></span></span>';			
+			document.getElementById('goodGuyName').innerHTML += ' <span class="badge antiBadge sugarRushBadge" id="sugarRushBuff" onmouseover="tooltip(\'Sugar Rush\', \'customText\', event, sugarRush.tooltipText())" onmouseout="tooltip(\'hide\')"><span class="' + this.icon + '"></span></span>';
 			return;
 		}
 		elem.style.display = 'inline-block';
@@ -16822,7 +16822,7 @@ function formatSecondsForDescriptions(number){
 		var s = (minutes > 1) ? "s" : "";
 		text = minutes + " min" + s;
 	}
-	return text;	
+	return text;
 }
 
 function formatSecondsForZoneTime(number){
@@ -16840,7 +16840,7 @@ function formatSecondsForZoneTime(number){
 		if (hours < 10) hours = "0" + hours;
 		text = hours + ":" + minutes + ":" + seconds;
 	}
-	return text;	
+	return text;
 }
 
 function getMinutesThisPortal(){
@@ -17099,7 +17099,7 @@ function saveAutoStructureConfig(){
 		if (nurseryZoneElem !== null && nurseryZoneElem.value >= 1 && !isNumberBad(nurseryZoneElem.value)){
 			setting.NurseryZones = nurseryZoneElem.value;
 		}
-		else if (typeof setting.NurseryZones !== 'undefined') 
+		else if (typeof setting.NurseryZones !== 'undefined')
 			delete setting.NurseryZones;
 	}
 	cancelTooltip();
@@ -17241,7 +17241,7 @@ function buyAutoEquip(){
 		var wantToBuy = calculateMaxAfford(game.equipment[item], false, true, false, buyMax, settingValue / 100);
 		if (buyMax != 0 && (wantToBuy > (buyMax - equip.level))) wantToBuy = buyMax - equip.level;
 		if (wantToBuy > 0) buyEquipment(item, true, true, wantToBuy);
-	}	
+	}
 }
 
 
@@ -17421,7 +17421,7 @@ var archoGolden = {
 		cancelTooltip();
 	},
 	getNext: function(){
-		if (game.global.guString == "") return false;		
+		if (game.global.guString == "") return false;
 		var defs = this.getDefs();
 		var split = game.global.guString.split(',');
 		var done = {};
@@ -17547,7 +17547,7 @@ function getCheapestPrestigeUpgrade(upgradeArray) {
 			shieldCheck = true;
 			shieldCost = thisCost;
 		}
-		
+
 		if (cheapest[1] == -1 || thisCost < cheapest[1]) cheapest = [upgradeArray[x], thisCost];
 	}
 	if (cheapest[0] && cheapest[0] != 'Supershield' && shieldCheck && shieldCost != -1 && game.resources.metal.owned < cheapest[1]) cheapest = ['Supershield', shieldCost];
@@ -17569,7 +17569,7 @@ function autoBuyUpgrade(item){
 var Fluffy = {
 	firstLevel: 1000,
 	getFirstLevel: function () {
-		var prestigeRequire = Math.pow(this.prestigeExpModifier, this.getCurrentPrestige());	
+		var prestigeRequire = Math.pow(this.prestigeExpModifier, this.getCurrentPrestige());
 		return this.firstLevel * prestigeRequire;
 	},
 	growth: 4,
@@ -17876,7 +17876,7 @@ var Fluffy = {
 			bottomElem.innerHTML = fluffyTip[1];
 			if (!justOnce) setTimeout(Fluffy.refreshTooltip, 1000);
 		}
-		
+
 	},
 	checkAndRunVoidance: function() {
 		if (!this.isRewardActive('voidance')) return;
@@ -17904,7 +17904,7 @@ var Fluffy = {
 				elem.innerHTML = "";
 				return;
 			case "daily":
-				var text = 'Applies when running a Daily Challenge, and matches the extra ' + heliumOrRadon() + ' from your Daily.' 
+				var text = 'Applies when running a Daily Challenge, and matches the extra ' + heliumOrRadon() + ' from your Daily.'
 				text += ((game.global.challengeActive == "Daily") ? ' Currently ' + prettify(1 + (getDailyHeliumValue(countDailyWeight()) / 100)) + '.' : ' Currently 1.');
 				text += " Does not apply to Best " + Fluffy.getName() + " Exp."
 				elem.innerHTML = text;
@@ -17976,9 +17976,9 @@ var Fluffy = {
 		var prestigeRewardsList = this.getPrestigeRewardList();
 		var calculatedPrestige = this.getCurrentPrestige();
 		var name = this.getName();
-		if (game.talents.fluffyAbility.purchased) 
+		if (game.talents.fluffyAbility.purchased)
 			calculatedPrestige++;
-		if (calculatedPrestige > prestigeRewardsList.length) 
+		if (calculatedPrestige > prestigeRewardsList.length)
 			calculatedPrestige = prestigeRewardsList.length + 1;
 
 		if (showCruffys){
@@ -18024,7 +18024,7 @@ var Fluffy = {
 					topText += "<br/>- " + Fluffy.getFluff();
 				}
 			}
-			
+
 		}
 		else topText += "- " + Fluffy.getFluff();
 		topText += "</br>";
@@ -18055,7 +18055,7 @@ var Fluffy = {
 			if (game.global.universe == 2 && u2Mutations.tree.Scruffy.purchased) fluffFormula += ' * <span class="fluffFormMutators" onmouseover="Fluffy.expBreakdown(\'mutators\')" onmouseout="Fluffy.expBreakdown(\'clear\')">Mutators</span>';
 			fluffFormula = fluffFormula.replace('Zone', '<span onmouseover="Fluffy.expBreakdown(\'zone\')" onmouseout="Fluffy.expBreakdown(\'clear\')" class="fluffFormZone">Zone</span>');
 			fluffFormula = fluffFormula.replace('Cunning', '<span onmouseover="Fluffy.expBreakdown(\'cunning\')" onmouseout="Fluffy.expBreakdown(\'clear\')" class="fluffFormCunning">Cunning</span>')
-			fluffFormula = fluffFormula.replace('Curious', '<span onmouseover="Fluffy.expBreakdown(\'curious\')" onmouseout="Fluffy.expBreakdown(\'clear\')" class="fluffFormCurious">Curious</span>')			
+			fluffFormula = fluffFormula.replace('Curious', '<span onmouseover="Fluffy.expBreakdown(\'curious\')" onmouseout="Fluffy.expBreakdown(\'clear\')" class="fluffFormCurious">Curious</span>')
 			topText += fluffFormula;
 		}
 		if (calculatedPrestige > 0 && Fluffy.currentLevel < 10 && !(calculatedPrestige == 1 && game.talents.fluffyAbility.purchased))
@@ -18066,7 +18066,7 @@ var Fluffy = {
 		for (var x = 0; x < rewardsList.length; x++){
 			var highlighted;
 			// if (showCruffys) highlighted = (cruffysLevel >= (x + 1));
-			// else 
+			// else
 			highlighted = ((fluffyInfo[0] + calculatedPrestige) >= (x + 1));
 			bottomText += (highlighted) ? "<tr class='fluffyRowComplete'>" : "<tr>";
 			if (savedLevel < x && calculatedPrestige == 0 && game.global.universe == 1)
@@ -18247,7 +18247,7 @@ var Fluffy = {
 		SADailies: {
 			get description(){
 				var cleared = autoBattle.maxEnemyLevel - 1;
-				return "Your Trimps gain +4% Attack and +0.25% Crit Chance per Spire Assault level cleared while on a Daily Challenge. You have cleared " + cleared + " SA levels, granting +" + prettify((this.attackMod() - 1) * 100) + "% Attack and +" + prettify(this.critChance() * 100) + "% Crit Chance on Daily Challenges." 
+				return "Your Trimps gain +4% Attack and +0.25% Crit Chance per Spire Assault level cleared while on a Daily Challenge. You have cleared " + cleared + " SA levels, granting +" + prettify((this.attackMod() - 1) * 100) + "% Attack and +" + prettify(this.critChance() * 100) + "% Crit Chance on Daily Challenges."
 			},
 			critChance: function(){
 				return (0.0025 * (autoBattle.maxEnemyLevel - 1));
@@ -18867,7 +18867,7 @@ function gameTimeout() {
 	}
 	var now = new Date().getTime();
 	//4432
-	if ((now - game.global.start - game.global.time) > 3600000){	
+	if ((now - game.global.start - game.global.time) > 3600000){
 		checkOfflineProgress();
 		game.global.start = now;
 		game.global.time = 0;
@@ -18936,7 +18936,7 @@ function formatSecondsAsClock(timeSince, maxOutputs){
 function preventZoom(elem){
 	elem.addEventListener("wheel", zoomShortcut); //add the event
 }
-  
+
 function zoomShortcut(e){
 	if(e.ctrlKey){
 		event.preventDefault();
@@ -19022,7 +19022,7 @@ document.addEventListener('keydown', function (e) {
 			}
 			e.preventDefault();
 			break;
-		
+
 		case 192: // `
 		case 48: //0
 		case 96: //num0
@@ -19137,7 +19137,7 @@ document.addEventListener('keydown', function (e) {
 				fightManual();
 			}
 			break;
-		case 73: 
+		case 73:
 			//i for spIre Assault
 			if (game.global.highestRadonLevelCleared >= 74){
 				if (game.global.lockTooltip && lastTooltipTitle == "Spire Assault") cancelTooltip();
@@ -19178,11 +19178,11 @@ document.addEventListener('keydown', function (e) {
 
 			}
 			break;
-		case 38: 
+		case 38:
 			//Up arrow for map levels
 			mapLevelHotkey(true);
 			break;
-		case 40: 
+		case 40:
 			//Down arrow for map levels
 			mapLevelHotkey(false);
 			break;
@@ -19236,7 +19236,7 @@ document.addEventListener('keyup', function(e) {
 		toggleGeneticistassist(true);
 		if (game.global.buyTab == "nature")
 			updateNatureInfoSpans();
-		if (game.global.buyTab == "talents") 
+		if (game.global.buyTab == "talents")
 			displayTalents();
 	}
