+++ src/entities/main-screen/screen.tsx
@@ -1,7 +1,7 @@
 import bgImage from "#assets/images/bg2.png";
 import { BottomRow } from "./bottom-row";
 import { Settings } from "./settings";
-import { TopRow } from "./top-row";
+import { TopRow } from "./top-row/top-row";
 
 import styles from "./screen.module.scss"
 
@@ -19,8 +19,6 @@ export function MainScreen() {
         <TopRow />
 
         <BottomRow />
-
-
       </div>
     </div>
   );
