--- src/entities/main-screen/top-row.module.scss
@@ -1,26 +1,0 @@
-.block {
-  display: flex;
-  flex-flow: row nowrap;
-  justify-content: space-between;
-  min-height: 25em;
-}
-
-.logs {
-  max-width: 50em;
-  height: 100%;
-}
-
-.resources {
-  display: flex;
-  flex-flow: row nowrap;
-  gap: 1em;
-  height: 100%;
-}
-
-.miscelaneous {
-
-}
-
-.trimps {
-  height: 100%;
-}
