+++ src/entities/main-screen/top-row/logs.module.scss
@@ -1,0 +1,11 @@
+.block {
+  position: relative;
+  max-width: 50em;
+  height: 100%;
+  color: white;
+  background: rgba(121, 121, 123, 0.5);
+  border: 2px solid white;
+}
+
+.achievement {
+}
