+++ src/entities/main-screen/top-row/miscelaneous.module.scss
@@ -1,0 +1,3 @@
+.block {
+
+}
