+++ src/entities/main-screen/top-row/resources.module.scss
@@ -1,0 +1,6 @@
+.block {
+  display: flex;
+  flex-flow: row nowrap;
+  gap: 1em;
+  height: 100%;
+}
