+++ src/entities/main-screen/top-row/top-row.module.scss
@@ -1,0 +1,6 @@
+.block {
+  display: flex;
+  flex-flow: row nowrap;
+  justify-content: space-between;
+  min-height: 25em;
+}
