+++ src/entities/main-screen/top-row/trimps.module.scss
@@ -1,0 +1,3 @@
+.block {
+  height: 100%;
+}
