+++ src/styles/original/style.scss
@@ -602,14 +602,6 @@ body {
   font-size: 1.2em;
 }
 
-#logContainer {
-  border: 2px solid white;
-  height: 100%;
-  width: 100%;
-  background: rgba(121, 121, 123, 0.5);
-  color: white;
-}
-
 #buyCol {
   height: 100%;
   padding-right: 0.25vw;
@@ -1311,7 +1303,6 @@ body {
   width: 100%;
   overflow-y: scroll;
   padding-left: 5px;
-  padding-top: 2px;
   padding-right: 5px;
 }
 
@@ -7113,14 +7104,13 @@ sup {
   border-color: #3f3f3f;
 }
 
+// @TODO: move into module after dealing with js replacing classnames
 #achievementTracker {
   position: absolute;
-  top: 11.5%;
-  width: calc(40.5vw - 15px);
-  height: 4.8vw;
+  width: 100%;
   z-index: 2;
   color: #3f3f3f;
-  font-size: 1vw;
+  font-size: 1.5em;
   text-align: center;
   border-radius: 25px;
   border: 4px outset #536373;
